requirements.txt
alembic==1.7.4
amqp==5.0.6
anyio==3.3.2
asgiref==3.4.1
astroid==2.8.2
billiard==3.6.4.0
black==21.9b0
Brotli==1.0.9
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cssselect2==0.4.1
Django==3.2.8
django-braces==1.14.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-embed-video==1.4.0
django-environ==0.7.0
django-extensions==3.1.3
django-filter==21.1
django-grappelli==2.15.3
django-js-asset==1.2.2
django-mssql-backend==2.8.1
django-redis==5.0.0
django-taggit==1.5.1
django-timezone-field==4.2.1
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
ecdsa==0.17.0
fastapi==0.70.0
flake8==3.9.2
flower==1.0.0
fonttools==4.27.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httptools==0.3.0
humanize==3.12.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
lazy-object-proxy==1.6.0
loguru==0.5.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
passlib==1.7.4
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydantic-sqlalchemy==0.0.9
pydyf==0.1.1
pyflakes==2.3.1
pylint==2.11.1
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyphen==0.11.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.3
redis==3.5.3
regex==2021.9.30
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
sorl-thumbnail==12.7.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
starlette==0.16.0
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
Werkzeug==2.0.2
win32-setctime==1.0.3
wrapt==1.12.1
zipp==3.6.0
zopfli==0.1.8
